requirements/dev.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements/dev.in
#
asgiref==3.5.2
    # via
    #   -r requirements/base.txt
    #   django
attrs==22.1.0
    # via
    #   -r requirements/base.txt
    #   pytest
bandit==1.7.4
    # via -r requirements/dev.in
black==22.1.0
    # via blue
blue==0.9.1
    # via -r requirements/dev.in
boto3==1.24.48
    # via -r requirements/base.txt
botocore==1.27.48
    # via
    #   -r requirements/base.txt
    #   boto3
    #   s3transfer
cachecontrol[filecache]==0.12.11
    # via pip-audit
certifi==2022.6.15
    # via requests
charset-normalizer==2.1.0
    # via requests
click==8.1.3
    # via black
commonmark==0.9.1
    # via rich
coverage[toml]==6.4.3
    # via
    #   -r requirements/base.txt
    #   pytest-cov
cyclonedx-python-lib==2.7.1
    # via pip-audit
dj-database-url==1.0.0
    # via -r requirements/base.txt
django==4.1
    # via
    #   -r requirements/base.txt
    #   dj-database-url
    #   django-ckeditor
    #   django-extensions
    #   django-js-asset
    #   django-storages
django-ckeditor==6.4.2
    # via -r requirements/base.txt
django-extensions==3.2.0
    # via -r requirements/dev.in
django-js-asset==2.0.0
    # via
    #   -r requirements/base.txt
    #   django-ckeditor
django-storages==1.13.1
    # via -r requirements/base.txt
flake8==4.0.1
    # via
    #   -r requirements/dev.in
    #   blue
    #   flake8-debugger
    #   pep8-naming
flake8-debugger==4.1.2
    # via -r requirements/dev.in
flake8-plugin-utils==1.3.2
    # via flake8-pytest-style
flake8-pytest-style==1.6.0
    # via -r requirements/dev.in
gitdb==4.0.9
    # via gitpython
gitpython==3.1.27
    # via bandit
html5lib==1.1
    # via pip-audit
idna==3.3
    # via requests
iniconfig==1.1.1
    # via
    #   -r requirements/base.txt
    #   pytest
isort==5.10.1
    # via -r requirements/dev.in
jmespath==1.0.1
    # via
    #   -r requirements/base.txt
    #   boto3
    #   botocore
lockfile==0.12.2
    # via cachecontrol
mccabe==0.6.1
    # via flake8
msgpack==1.0.4
    # via cachecontrol
mypy==0.971
    # via -r requirements/dev.in
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
packageurl-python==0.10.1
    # via cyclonedx-python-lib
packaging==21.3
    # via
    #   -r requirements/base.txt
    #   pip-audit
    #   pip-requirements-parser
    #   pytest
pathspec==0.9.0
    # via black
pbr==5.9.0
    # via stevedore
pep8-naming==0.13.1
    # via -r requirements/dev.in
pillow==9.2.0
    # via -r requirements/base.txt
pip-api==0.0.30
    # via pip-audit
pip-audit==2.4.3
    # via -r requirements/dev.in
pip-requirements-parser==31.2.0
    # via pip-audit
platformdirs==2.5.2
    # via black
pluggy==1.0.0
    # via
    #   -r requirements/base.txt
    #   pytest
psycopg2-binary==2.9.3
    # via -r requirements/base.txt
py==1.11.0
    # via
    #   -r requirements/base.txt
    #   pytest
pycodestyle==2.8.0
    # via
    #   flake8
    #   flake8-debugger
pyflakes==2.4.0
    # via flake8
pygments==2.12.0
    # via rich
pyparsing==3.0.9
    # via
    #   -r requirements/base.txt
    #   packaging
pytest==7.1.2
    # via
    #   -r requirements/base.txt
    #   pytest-cov
    #   pytest-django
pytest-cov==3.0.0
    # via -r requirements/base.txt
pytest-django==4.5.2
    # via -r requirements/base.txt
python-dateutil==2.8.2
    # via
    #   -r requirements/base.txt
    #   botocore
python-decouple==3.6
    # via -r requirements/base.txt
pyupgrade==2.37.3
    # via -r requirements/dev.in
pyyaml==6.0
    # via bandit
requests==2.28.1
    # via cachecontrol
resolvelib==0.8.1
    # via pip-audit
rich==12.5.1
    # via pip-audit
s3transfer==0.6.0
    # via
    #   -r requirements/base.txt
    #   boto3
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   html5lib
    #   python-dateutil
smmap==5.0.0
    # via gitdb
sortedcontainers==2.4.0
    # via cyclonedx-python-lib
sqlparse==0.4.2
    # via
    #   -r requirements/base.txt
    #   django
stevedore==4.0.0
    # via bandit
tokenize-rt==4.2.1
    # via pyupgrade
toml==0.10.2
    # via
    #   cyclonedx-python-lib
    #   pip-audit
tomli==2.0.1
    # via
    #   -r requirements/base.txt
    #   black
    #   coverage
    #   mypy
    #   pytest
typing-extensions==4.3.0
    # via mypy
urllib3==1.26.11
    # via
    #   -r requirements/base.txt
    #   botocore
    #   requests
webencodings==0.5.1
    # via html5lib

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
